Fix incorrect condition that is always false.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11954 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ server/shop.c
@@ -1184,7 +1184,7 @@ static double shop_specialisation_ratio(const object *item, const mapstruct *map
         LOG(llevError, "shop_specialisation_ratio: passed a NULL item for map %s\n", map->path);
         return 0;
     }
-    if (item->type == -1) {
+    if (item->type == (uint8)-1) {
         LOG(llevError, "shop_specialisation_ratio: passed an item with an invalid type\n");
         /*
          * I'm not really sure what the /right/ thing to do here is,
